Fix: comment bugs

+++ frontend/src/app/comments/comment/comment.component.ts
@@ -1,4 +1,4 @@
-import { Component, EventEmitter, Input, OnChanges, Output, SimpleChanges} from '@angular/core';
+import { Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges} from '@angular/core';
 import { Comment } from '../models/comment.model';
 import TimeAgo from 'javascript-time-ago';
 import en from 'javascript-time-ago/locale/en';
@@ -12,12 +12,13 @@ TimeAgo.addDefaultLocale(en);
 	templateUrl: './comment.component.html',
 	styleUrls: ['./comment.component.css']
 })
-export class CommentComponent implements OnChanges{
+export class CommentComponent implements OnChanges, OnInit{
 
 	public timeFormater : any;
 	public commentModified : boolean = false;
 	public editingMode : boolean = false;
 	public editForm : FormGroup;
+	public ownsComment : boolean = false;
 
 	@Output('commentDeleted')
  	public emitCommentDeleted : EventEmitter<string> = new EventEmitter<string>();
@@ -41,6 +42,9 @@ export class CommentComponent implements OnChanges{
 		"",
 		"");
 
+	@Input("userId")
+	public userId : string = "";
+
 	constructor(private commentService : CommentService, private formBuilder: FormBuilder) {
 		this.timeFormater = new TimeAgo('en-US');
 		
@@ -50,7 +54,10 @@ export class CommentComponent implements OnChanges{
 
 		this.editForm = this.formBuilder.group({
 			content : [this.comment.content, [Validators.required, Validators.minLength(2)]]
-		})
+		});
+	}
+	ngOnInit(): void {
+		this.ownsComment = this.userId == this.comment.authorId;
 	}
 	ngOnChanges(changes: SimpleChanges): void {
 		this.editForm.controls["content"].setValue(this.comment.content);
@@ -62,18 +69,23 @@ export class CommentComponent implements OnChanges{
 	}
 
 	public upvote() {
-		this.commentService.upvote(this.comment._id).subscribe((res: {currentVoteCount : number}) => {
+		this.commentService.upvote(this.comment._id, this.userId).subscribe((res: {currentVoteCount : number}) => {
 			this.comment.votes = res.currentVoteCount;
 		});
 	}
 
 	public downvote() {
-		this.commentService.downvote(this.comment._id).subscribe((res : {currentVoteCount : number}) => {
+		this.commentService.downvote(this.comment._id, this.userId).subscribe((res : {currentVoteCount : number}) => {
 			this.comment.votes = res.currentVoteCount;
 		});
 	}
 
 	public deleteComment() {
+		if (this.userId != this.comment.authorId) {
+			window.alert("You can not delete comment that is not yours!");
+			return;
+		} 
+
 		this.commentService.deleteComment(this.comment._id).subscribe((res: any) => {});
 		this.emitCommentDeleted.emit(this.comment._id);
 	}
